% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{add_bound}
\alias{add_bound}
\title{Add a Sensitivity Bound to a \code{"sensana"} object}
\usage{
add_bound(
  sa,
  arrow = c("UD", "UY", "ZU", "ZY"),
  kind = c("direct", "comparative", "comparative-d"),
  lb = -1,
  ub = 1,
  b = 1,
  I = NULL,
  J = NULL,
  name = "",
  print_warning = TRUE
)
}
\arguments{
\item{sa}{Object of the class \code{sensana}.}

\item{arrow}{String indicating the type of arrow to add.
One of \code{"UD"}, \code{"UY"}, \code{"ZU"}, or \code{"ZY"}.}

\item{kind}{String specifying the type of the bound.
One of \code{"direct"}, \code{"comparative"}, or \code{"comparative-d"}.}

\item{lb}{Numeric value specifying the lower bound. Default is \code{-1}.}

\item{ub}{Numeric value specifying the upper bound. Default is \code{1}.}

\item{b}{Positive numeric value used for comparative bounds. Default is \code{1}.}

\item{I}{Character vector of covariate indices/names used for comparative bounds;
subset of \code{sa$indep_x}. Default is \code{NULL}.}

\item{J}{Character vector of covariate indices/names used for comparative bounds;
subset of \code{sa$indep_x}. Default is \code{NULL}.}

\item{name}{Character string giving a name to the bound. Default is an empty string.}

\item{print_warning}{Logical; if \code{TRUE}, a warning will be printed if the bound
is does not have an effect on the estimation of the partially identified range.
Default is \code{TRUE}.}
}
\value{
An updated \code{sensana} object that includes the added bound in its
\code{bounds} data frame.
}
\description{
This method adds a new constraint / bound to the sensitivity model specified
in \code{sa}. Users can specify bounds on the effects of \eqn{U} on \eqn{D} (\code{"UD"}),
\eqn{U} on \eqn{Y} (\code{"UY"}), \eqn{Z} on \eqn{U} (\code{"ZU"}) or
\eqn{Z} on \eqn{Y} (\code{"ZY"}). There are 3 different availble kinds of bounds:
\code{"direct"}, \code{"comparative"} and \code{"comparative-d"}.
(The latter is only suitable for \code{"UY"}.)
}
\details{
Arguments not pertaining to the specified \code{arrow} and \code{kind} are ignored.
If a comparative bound on \code{"ZU"} or \code{"ZY"} is given, \code{J} must be a single
string; moreover, it will be ignored if \code{sa$indep_x} contains only one entry.

For the interpretation of the bounds, see the article cited below.
}
\examples{
set.seed(123)
u <- rnorm(20)
xt <- -0.25*u + rnorm(20)
xp <- 0.5 * xt + rnorm(20)
z <- 0.25 * xt - xp + u + rnorm(20)
d <- 0.5 * xt + 0.5 * xp + u + 2*z + rnorm(20)
y <- d + 2*xp - xt + z + u + rnorm(20)
sa <- sensana(y = y, d = d, x = data.frame(xt = xt, xp = xp), z = z,
              dep_x = "xt", indep_x = "xp", alpha = 0.05, quantile = "t")

## Direct bounds
sa <- add_bound(sa, arrow = "UD", kind = "direct", lb = -0.3, ub = 0.3)
sa <- add_bound(sa, arrow = "ZY", kind = "direct", lb = -0.3, ub = 0.3)

## Comparative bounds
sa <- add_bound(sa, arrow = "UD", kind = "comparative",
                b = 0.2, J = "xp", I = NULL)
sa <- add_bound(sa, arrow = "UY", kind = "comparative-d",
                b = 0.3, J = "xp", I = NULL)
sa <- add_bound(sa, arrow = "ZU", kind = "comparative",
                b = 0.4, J = "xp")

}
\references{
Freidling T, Zhao Q (2024). “Optimization-based Sensitivity Analysis for
Unmeasured Confounding using Partial
Correlations.” \emph{arXiv preprint arXiv:2301.00040v3}.
}
