% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{feasible_grid}
\alias{feasible_grid}
\title{Compute a grid of feasible points.}
\usage{
feasible_grid(y, d, xt, xp, z, bounds, grid_specs, full_grid, eps = 0.001)
}
\arguments{
\item{y}{A numeric vector of \emph{centered} outcomes.}

\item{d}{A numeric vector of \emph{centered} treatments.}

\item{xt}{A matrix of \emph{centered} not conditionally independent covariates.}

\item{xp}{A matrix of \emph{centered} conditionally independent covariates.}

\item{z}{A numeric vector of \emph{centered} instruments.}

\item{bounds}{A data frame containing bounds according to the format used by
the class \code{"sensana"}.}

\item{grid_specs}{A named list of the three numeric values \code{N1}, \code{N2} and \code{N5} specifying
the number of points considered in the grid search for each of the three dimensions.}

\item{full_grid}{Logical; if \code{TRUE}, provides the full grid of
\eqn{(\psi_1,\psi_2)} values. If \code{FALSE}, provides only the smallest and
largest \eqn{\psi_2} value for each \eqn{\psi_1}. (This suffices for estimating
the partially identified range.)}

\item{eps}{A small numeric number to bound sensitivity parameters away from \code{1}
and \code{-1}. Default is \code{0.001}.}
}
\value{
A list containing two elements:
\describe{
\item{p1_seq  }{A numeric vector of length \code{N1} containing the grid points for
the sensitivity parameter \eqn{\psi_1}.}
\item{p2_mat  }{A numeric matrix with dimension \code{(N1, N2)} if \code{full_grid == TRUE} and
\code{(N1, 2)} otherwise. Row \code{i} of the matrix contains the grid points for \eqn{\psi_2}
belonging to the \code{i}-th grid point of \eqn{\psi_1}, i.e., the \code{i}-th entry in \code{p1_seq}.}
}
}
\description{
Constructs a grid of \eqn{(\psi_1,\psi_2)} points that are (approximately) contained in the
sensitivity model \eqn{\Psi} which is defined via the user-specified bounds.
This method should only be used to directly access the grid-search algorithm.
For inference, see \code{\link[=pir]{pir()}} and \code{\link[=sensint]{sensint()}}; for visualizations, see \code{\link[=b_contours]{b_contours()}}
and \code{\link[=r_contours]{r_contours()}}.
}
\examples{
set.seed(123)
u <- rnorm(20)
xt <- -0.25*u + rnorm(20)
xp <- 0.5 * xt + rnorm(20)
z <- 0.25 * xt - xp + u + rnorm(20)
d <- 0.5 * xt + 0.5 * xp + u + 2*z + rnorm(20)
y <- d + 2*xp - xt + z + u + rnorm(20)
sa <- sensana(y = y, d = d, x = data.frame(xt = xt, xp = xp), z = z,
              dep_x = "xt", indep_x = "xp", alpha = 0.05, quantile = "t")

sa <- add_bound(sa, arrow = "UD", kind = "direct", lb = -0.3, ub = 0.3)
sa <- add_bound(sa, arrow = "ZY", kind = "direct", lb = -0.3, ub = 0.3)
                
grid <- feasible_grid(y = sa$y, d = sa$d, xt = sa$xt, xp = sa$xp, z = sa$z,
                      bounds = sa$bounds,
                      grid_specs = list(N1 = 100, N2 = 100, N5 = 100),
                      full_grid = TRUE, eps = 0.001)

}
