% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{pir}
\alias{pir}
\title{Compute the Partially Identified Range (PIR)}
\usage{
pir(sa, grid_specs = list(N1 = 100, N2 = 100, N5 = 100), eps = 0.001)
}
\arguments{
\item{sa}{An object of the class \code{sensana}.}

\item{grid_specs}{A named list of the three numeric values \code{N1}, \code{N2} and \code{N5} specifying
the number of points considered in the grid search for each of the three dimensions.}

\item{eps}{A small numeric number to bound sensitivity parameters away from \code{1}
and \code{-1}. Default is \code{0.001}.}
}
\value{
A numeric vector of length \code{2}, the estimated partially identified range.
}
\description{
Computes the PIR by finding the smallest and largest possible value of the
causal effect according to the specified sensitivity model. To solve this
optimization problem a tailored grid search algorithm is used.
}
\examples{
set.seed(123)
u <- rnorm(20)
xt <- -0.25*u + rnorm(20)
xp <- 0.5 * xt + rnorm(20)
z <- 0.25 * xt - xp + u + rnorm(20)
d <- 0.5 * xt + 0.5 * xp + u + 2*z + rnorm(20)
y <- d + 2*xp - xt + z + u + rnorm(20)
sa <- sensana(y = y, d = d, x = data.frame(xt = xt, xp = xp), z = z,
              dep_x = "xt", indep_x = "xp", alpha = 0.05, quantile = "t")

sa <- add_bound(sa, arrow = "UD", kind = "direct", lb = -0.3, ub = 0.3)
sa <- add_bound(sa, arrow = "ZY", kind = "direct", lb = -0.3, ub = 0.3)
                
pir_res <- pir(sa, grid_specs = list(N1 = 100, N2 = 100, N5 = 100))

}
