% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensana.R
\name{print.sensana}
\alias{print.sensana}
\title{Print Method for \code{sensana} objects}
\usage{
\method{print}{sensana}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{"sensana"}.}

\item{digits}{The number of significant digits to display.}

\item{...}{Additional arguments passed to or from other methods (ignored here).}
}
\value{
Invisibly returns \code{x}, the original object.
}
\description{
Print Method for \code{sensana} objects
}
\examples{
x <- rnorm(20)
z <- 0.5 * x + rnorm(20)
d <- z - 0.5 * x + rnorm(20)
y <- d + x + rnorm(20)
sa <- sensana(y = y, d = d, x = data.frame(x = x), z = z,
              indep_x = "x", dep_x = NULL, alpha = 0.05, quantile = "t")
print(sa)
}
