% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{print.sensint}
\alias{print.sensint}
\title{Print Method for \code{sensint} objects}
\usage{
\method{print}{sensint}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{"sensint"}.}

\item{digits}{The number of significant digits to display.}

\item{...}{Additional arguments passed to or from other methods (ignored here).}
}
\value{
Invisibly returns \code{x}, the original object.
}
\description{
Print Method for \code{sensint} objects
}
\examples{
set.seed(123)
u <- rnorm(20)
xt <- -0.25*u + rnorm(20)
xp <- 0.5 * xt + rnorm(20)
z <- 0.25 * xt - xp + u + rnorm(20)
d <- 0.5 * xt + 0.5 * xp + u + 2*z + rnorm(20)
y <- d + 2*xp - xt + z + u + rnorm(20)
sa <- sensana(y = y, d = d, x = data.frame(xt = xt, xp = xp), z = z,
              dep_x = "xt", indep_x = "xp", alpha = 0.05, quantile = "t")

sa <- add_bound(sa, arrow = "UD", kind = "direct", lb = -0.3, ub = 0.3)
sa <- add_bound(sa, arrow = "ZY", kind = "direct", lb = -0.3, ub = 0.3)
                
sensint_obj <- sensint(sa, alpha = 0.05)
print(sensint_obj)

}
