% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-contours.R
\name{r_contours}
\alias{r_contours}
\title{Generate an R-Contour Plot}
\usage{
r_contours(
  sa,
  val_interest = 0,
  comparison_ind = NULL,
  comparison = c("comp", "comp-d", "informal"),
  iv_lines = FALSE,
  bw = FALSE,
  grid_specs = list(N1 = 100, N2 = 100, N5 = 100),
  eps = 0.001,
  n_breaks = 15
)
}
\arguments{
\item{sa}{An object of the class \code{sensana}.}

\item{val_interest}{Numeric; the value of interest for the user; Default is \code{0}.}

\item{comparison_ind}{A named list specifying which covariates at which "predictive
strength" are used for the comparison points. For example, \code{list(x1 = c(1,3))}
signifies that one comparison point for the covariate named \code{x1} and one comparison
point for the covariate \code{x1} if it were 3 times as strong will be plotted.
Default is \code{NULL} for no comparison points.}

\item{comparison}{A character string indicating the type of comparison. Must be one of:
\code{"comp"} (default), \code{"comp-d"}, or \code{"informal"}.}

\item{iv_lines}{Logical. If \code{TRUE}, IV lines are plotted. Default is \code{FALSE}.}

\item{bw}{Logical; whether to generate a black-and-white plot. Default is \code{FALSE}.}

\item{grid_specs}{A named list of the three numeric values \code{N1}, \code{N2} and \code{N5} specifying
the number of points considered in the grid search for each of the three dimensions.}

\item{eps}{A small numeric number to bound sensitivity parameters away from \code{1}
and \code{-1}. Default is \code{0.001}.}

\item{n_breaks}{Integer specifying the number of envisaged contour levels.
Default is \code{15}.}
}
\value{
An object of the class \code{ggplot} that can be printed.
}
\description{
For a given sensitivity model, this method plots the estimated causal effect
as a function of the sensitivity parameters \eqn{\psi_1 = R_{D\sim U\vert X,Z}}
and \eqn{\psi_2 = R_{Y\sim U\vert X,Z,D}} for all feasible values.
Users can add comparison points based on the partial correlations of the covariates
to provide context for plausible values of the sensitivity parameters. If the
instrumental variable is valid, \eqn{R_{D\sim U\vert X,Z}} and \eqn{R_{Y\sim U\vert X,Z,D}}
fulfill an equation which defines lines in the \eqn{(\psi_1,\psi_2)}-plane. Users
can choose to plot these lines as well.
}
\examples{
set.seed(123) 
u <- rnorm(20)
xt <- -0.25*u + rnorm(20)
xp <- 0.5 * xt + rnorm(20)
z <- 0.25 * xt - xp + u + rnorm(20)
d <- 0.5 * xt + 0.5 * xp + u + 2*z + rnorm(20)
y <- d + 0.1*xp - xt + z + u + rnorm(20)
sa <- sensana(y = y, d = d, x = data.frame(xt = xt, xp = xp), z = z,
              dep_x = "xt", indep_x = "xp", alpha = 0.05, quantile = "t")

sa <- add_bound(sa, arrow = "UD", kind = "direct", lb = -0.3, ub = 0.3)
sa <- add_bound(sa, arrow = "ZY", kind = "direct", lb = -0.3, ub = 0.3)

comparison_ind <- list(xp = c(0.5, 1, 3))

pl <- r_contours(sa, val_interest = 0,
                 comparison_ind = comparison_ind, comparison = "comp",
                 iv_lines = FALSE)
# print(pl)

}
