% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-contours.R
\name{r_contours_data}
\alias{r_contours_data}
\title{Generate the Data for an R-Contour Plot}
\usage{
r_contours_data(
  sa,
  comparison_ind,
  comparison = c("comp", "comp-d", "informal"),
  iv_lines = FALSE,
  grid_specs = list(N1 = 100, N2 = 100, N5 = 100),
  eps = 0.001
)
}
\arguments{
\item{sa}{An object of the class \code{sensana}.}

\item{comparison_ind}{A named list specifying which covariates at which "predictive
strength" are used for the comparison points. For example, \code{list(x1 = c(1,3))}
signifies that one comparison point for the covariate named \code{x1} and one comparison
point for the covariate \code{x1} if it were 3 times as strong will be plotted.
Default is \code{NULL} for no comparison points.}

\item{comparison}{A character string indicating the type of comparison. Must be one of:
\code{"comp"} (default), \code{"comp-d"}, or \code{"informal"}.}

\item{iv_lines}{Logical. If \code{TRUE}, IV lines are plotted. Default is \code{FALSE}.}

\item{grid_specs}{A named list of the three numeric values \code{N1}, \code{N2} and \code{N5} specifying
the number of points considered in the grid search for each of the three dimensions.}

\item{eps}{A small numeric number to bound sensitivity parameters away from \code{1}
and \code{-1}. Default is \code{0.001}.}
}
\value{
A list containing two data frames and one list (or \code{NULL}):
\describe{
\item{df_plot}{A data frame with the columns \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}
specifying rectangles in the \eqn{(\psi_1,\psi_2)}-plane, \code{val}
the value of the causal effect in the rectangle, \code{feasible} indicating whether
the value of the sensitivity parameter is contained in the sensitivity model.
To be used with \code{\link[ggplot2:geom_tile]{ggplot2::geom_rect()}}.}
\item{df_cp}{A data frame containing the x- and y-coordinates of the comparison
points in the \eqn{(\psi_1,\psi_2)}-plane in the columns \code{cp_x} and \code{cp_y}.
A label for the comparison point is given in the column \code{cp_label}.}
\item{iv_lines_list}{A list of functions and numeric values needed to plot
the lines corresponding to valid instrumental variables. We recommend to
check the source code of the internal function \code{r_contours_plot()}.}
}
}
\description{
If users want to customize the plot generated by \code{\link[=r_contours]{r_contours()}}, they can use
this function to compute the data needed for the contour plot and then proceed
to plot it themselves.
}
\examples{
set.seed(123) 
u <- rnorm(20)
xt <- -0.25*u + rnorm(20)
xp <- 0.5 * xt + rnorm(20)
z <- 0.25 * xt - xp + u + rnorm(20)
d <- 0.5 * xt + 0.5 * xp + u + 2*z + rnorm(20)
y <- d + 0.1*xp - xt + z + u + rnorm(20)
sa <- sensana(y = y, d = d, x = data.frame(xt = xt, xp = xp), z = z,
              dep_x = "xt", indep_x = "xp", alpha = 0.05, quantile = "t")

sa <- add_bound(sa, arrow = "UD", kind = "direct", lb = -0.3, ub = 0.3)
sa <- add_bound(sa, arrow = "ZY", kind = "direct", lb = -0.3, ub = 0.3)

comparison_ind <- list(xp = c(0.5, 1, 3))

res <- r_contours(sa, comparison_ind = comparison_ind,
                  comparison = "comp", iv_lines = FALSE)
                  
# proceed to make customized plot using the data provided in res
}
