% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{remove_bound}
\alias{remove_bound}
\title{Remove a Sensitivity Bound from a \code{"sensana"} object}
\usage{
remove_bound(sa, name)
}
\arguments{
\item{sa}{An object of class \code{sensana}.}

\item{name}{A character string indicating the name of the bound to be removed.}
}
\value{
An updated \code{sensana} object with the specified bound removed.
}
\description{
Removes a previously added sensitivity bound from a \code{sensana} object by name.
}
\details{
To display the names of all the current bounds, print the \code{sensana} object or
directly access the \code{sa$bounds} data frame.
}
\examples{
u <- rnorm(20)
xt <- -0.25*u + rnorm(20)
xp <- 0.5 * xt + rnorm(20)
z <- 0.25 * xt - xp + u + rnorm(20)
d <- 0.5 * xt + 0.5 * xp + u + 2*z + rnorm(20)
y <- d + 2*xp - xt + z + u + rnorm(20)
sa <- sensana(y = y, d = d, x = data.frame(xt = xt, xp = xp), z = z,
              dep_x = "xt", indep_x = "xp", alpha = 0.05, quantile = "t")

## Add bound
sa <- add_bound(sa, arrow = "UD", kind = "direct", lb = -0.3, ub = 0.3, name = "example_bound")
## Remove bound
sa <- remove_bound(sa, name = "example_bound")

}
