% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensana.R
\name{sensana}
\alias{sensana}
\title{Sensitivity Analysis for Linear Regression and Instrumental Variable models}
\usage{
sensana(
  y,
  d,
  indep_x,
  dep_x,
  x = NULL,
  z = NULL,
  quantile = c("normal", "t"),
  alpha = 0.05
)
}
\arguments{
\item{y}{Numeric vector of outcomes.}

\item{d}{Numeric vector of treatments/endogeneous variables.}

\item{indep_x}{Character vector of names of cond. independent covariates. Default is \code{NULL}.}

\item{dep_x}{Character vector of names of remaining covariates. Default is \code{NULL}.}

\item{x}{Numeric matrix or data.frame of covariates. Default is \code{NULL}.}

\item{z}{Numeric vector of instruments. Default is \code{NULL}.}

\item{quantile}{Choice of quantile used in confidence interval: \code{"normal"}
(default) or \code{"t"}.}

\item{alpha}{Significance level for the confidence interval. Default is \code{0.05}.}
}
\value{
Object of the class \code{sensana}. This is a list
containing the following components:
\item{call}{The function call.}
\item{y}{The centered entries of the \code{y} vector.}
\item{d}{The centered entries of the \code{d} vector.}
\item{xt}{The centered matrix of covariates indexed by \code{dep_x}.}
\item{xp}{The centered matrix of covariates indexed by \code{indep_x}.}
\item{z}{The centered entries of the \code{z} vector.}
\item{beta_ols}{OLS estimate.}
\item{beta_tsls}{TSLS estimate.}
\item{confint_ols}{\code{1-alpha} OLS confidence interval.}
\item{confint_tsls}{\code{1-alpha} TSLS confidence interval.}
\item{se_ols}{Standard error of OLS estimator.}
\item{se_tsls}{Standard error of TSLS estimator.}
\item{bounds}{An empty data frame that stores bounds on the sensitivity
parameters. Users can add bounds via \code{\link[=add_bound]{add_bound()}}.}
\item{alpha}{Specified confidence level.}
\item{rbc}{Internal counter used for default names of bounds.}
}
\description{
\code{sensana()} does data pre-processing for the sensitivity analysis, computes the
OLS and TSLS estimates and confidence intervals and creates a
\code{sensana} object.
}
\details{
We split the covariates \eqn{X} into 2 groups: \eqn{\dot{X}}
and \eqn{\tilde{X}}. The arguments \code{indep_x} and \code{dep_x} contain the indices/names
of the respective covariates. For an unmeasured confounder \eqn{U}, we assume
that \eqn{\dot{X} \perp U \mid \tilde{X}} holds.

This method (and package) can be used even if there are no instrumental variables
or covariates. The absence of these is indicated by passing \code{NULL} as an argument
to the respective parameters.

To specify a sensitivity model, use the method \code{\link[=add_bound]{add_bound()}}. To conduct inference,
use the method \code{\link[=pir]{pir()}} to estimate the partially identified range and use the
method \code{\link[=sensint]{sensint()}} to estimate a sensitivity interval.

More details are provided in the article cited below.
}
\examples{
x <- rnorm(20)
z <- 0.5 * x + rnorm(20)
d <- z - 0.5 * x + rnorm(20)
y <- d + x + rnorm(20)
sa <- sensana(y = y, d = d, x = data.frame(x = x), z = z,
              indep_x = "x", dep_x = NULL, alpha = 0.05, quantile = "t")
              

}
\references{
Freidling T, Zhao Q (2024). “Optimization-based Sensitivity Analysis for
Unmeasured Confounding using Partial
Correlations.” \emph{arXiv preprint arXiv:2301.00040v3}.
}
